Segundo

+++ demo/src/main/java/com/example/demo/repositories/UserDAO.java
@@ -4,9 +4,9 @@ import com.example.demo.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface UserDAO extends JpaRepository<User, Long> {
-
+public interface UserDAO extends JpaRepository<User, Integer> {
 }
